requirements.txt
typer==0.12.3
PyYAML==6.0.2
rapidfuzz==3.9.7
